[NEW RBO] Split logical rename handling
